requirements.txt
typer
pyyaml
openai
xai-sdk
python-dotenv
apscheduler
fastapi
uvicorn
httpx
markdown
rapidfuzz
Pillow
pydantic-settings
chromadb
fastembed


pytest
coverage
